+++ UpgradeManagers/HeartOfGold.cs
@@ -1,4 +1,5 @@
-using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
 using BepInEx.Configuration;
 using HarmonyLib;
 using Newtonsoft.Json;
@@ -11,6 +12,7 @@ namespace SLRUpgradePack.UpgradeManagers;
 
 public class HeartOfGoldUpgrade : UpgradeBase<float> {
     public ConfigEntry<float> BaseHeartValue { get; protected set; }
+    internal Dictionary<string, GoldenHeart> GoldenHearts { get; set; } = new();
 
     public HeartOfGoldUpgrade(bool enabled, float upgradeAmount, bool exponential, float exponentialAmount,
                               ConfigFile config, AssetBundle assetBundle, float baseValue, float priceMultiplier) :
@@ -26,20 +28,26 @@ public class HeartOfGoldUpgrade : UpgradeBase<float> {
     internal override void InitUpgrade(PlayerAvatar player, int level) {
         base.InitUpgrade(player, level);
 
-        if (!player.healthGrab.gameObject.TryGetComponent<GoldenHeart>(out var heart)) {
-            heart = player.healthGrab.gameObject.AddComponent<GoldenHeart>();
-            heart.PlayerAvatarInstance = player;
-            heart.gameObject.AddComponent<PhotonView>();
-        }
+        if (GoldenHearts.TryGetValue(SemiFunc.PlayerGetSteamID(player), out var goldenHeart)) Object.Destroy(goldenHeart);
+
+        goldenHeart = new GameObject($"Golden Heart: {SemiFunc.PlayerGetName(player)}").AddComponent<GoldenHeart>();
+        goldenHeart.player = player;
+        GoldenHearts[SemiFunc.PlayerGetSteamID(player)] = goldenHeart;
     }
 }
 
 public class GoldenHeart : MonoBehaviour {
-    internal PlayerAvatar PlayerAvatarInstance { get; set; }
+    internal PlayerAvatar player { get; set; }
     internal bool Pause { get; set; }
     internal int lastLevel { get; set; } = -1;
     internal int lastHealth { get; set; } = -1;
+    internal ValuableObject? ValuableComponent { get; private set; }
+
     private PhotonView photonView;
+    private readonly FieldRef<PlayerHealth, int> _healthRef = FieldRefAccess<PlayerHealth, int>("health");
+    private readonly FieldRef<ValuableObject, PhysAttribute> _physAttributePresetRef = FieldRefAccess<ValuableObject, PhysAttribute>("physAttributePreset");
+    private readonly FieldRef<ValuableObject, bool> _discoveredRef = FieldRefAccess<ValuableObject, bool>("discovered");
+    private readonly FieldRef<ValuableObject, float> _dollarValueCurrentRef = FieldRefAccess<ValuableObject, float>("dollarValueCurrent");
 
     private void Start() {
         photonView = gameObject.GetComponent<PhotonView>();
@@ -52,9 +60,12 @@ public class GoldenHeart : MonoBehaviour {
 
     [PunRPC]
     private void DestroyOnlyMeRPC() {
-        SLRUpgradePack.Logger.LogInfo($"Components in GoldenHeart: {string.Join(',', (IEnumerable)GetComponents<Object>())}");
-        RoundDirector.instance.PhysGrabObjectRemove(GetComponent<ValuableObject>().GetComponent<PhysGrabObject>());
-        Destroy(GetComponent<ValuableObject>());
+        if (ValuableComponent == null) return;
+
+        SLRUpgradePack.Logger.LogInfo($"Components in GoldenHeart: {string.Join(',', GetComponents<Object>().Select(v => v.ToString()))}");
+        RoundDirector.instance.PhysGrabObjectRemove(ValuableComponent.GetComponent<PhysGrabObject>());
+        Destroy(ValuableComponent);
+        ValuableComponent = null;
     }
 
     public void CreateOnlyMe() {
@@ -64,52 +75,50 @@ public class GoldenHeart : MonoBehaviour {
 
     [PunRPC]
     private void CreateOnlyMeRPC() {
-        var healthRef = FieldRefAccess<PlayerHealth, int>("health");
         var heartOfGoldUpgrade = SLRUpgradePack.HeartOfGoldUpgradeInstance;
 
-        SLRUpgradePack.Logger.LogInfo($"Adding valuable component to player {SemiFunc.PlayerGetName(PlayerAvatarInstance)}");
+        SLRUpgradePack.Logger.LogInfo($"Adding valuable component to player {SemiFunc.PlayerGetName(player)}");
 
-        var valuableObject = gameObject.AddComponent<ValuableObject>();
+        ValuableComponent = player.healthGrab.gameObject.AddComponent<ValuableObject>();
 
-        SLRUpgradePack.Logger.LogInfo($"Valuable component {SemiFunc.PlayerGetName(PlayerAvatarInstance)} instantiated at {JsonConvert.SerializeObject(valuableObject.transform)}");
+        SLRUpgradePack.Logger.LogInfo($"Valuable component {SemiFunc.PlayerGetName(player)} instantiated at {JsonConvert.SerializeObject(ValuableComponent.transform)}");
 
-        valuableObject.valuePreset = ScriptableObject.CreateInstance<Value>();
-        valuableObject.valuePreset.valueMin =
-            valuableObject.valuePreset.valueMax =
-                heartOfGoldUpgrade.Calculate(healthRef.Invoke(PlayerAvatarInstance.playerHealth) * heartOfGoldUpgrade.BaseHeartValue.Value,
-                                             PlayerAvatarInstance, heartOfGoldUpgrade.UpgradeRegister.GetLevel(PlayerAvatarInstance));
+        ValuableComponent.valuePreset = ScriptableObject.CreateInstance<Value>();
+        ValuableComponent.valuePreset.valueMin =
+            ValuableComponent.valuePreset.valueMax =
+                heartOfGoldUpgrade.Calculate(_healthRef.Invoke(player.playerHealth) * heartOfGoldUpgrade.BaseHeartValue.Value,
+                                             player, heartOfGoldUpgrade.UpgradeRegister.GetLevel(player));
 
-        valuableObject.durabilityPreset = ScriptableObject.CreateInstance<Durability>();
-        valuableObject.durabilityPreset.durability = 999;
-        valuableObject.durabilityPreset.fragility = 0;
-        valuableObject.physAttributePreset = ScriptableObject.CreateInstance<PhysAttribute>();
-        valuableObject.transform.localScale = Vector3.zero;
-        valuableObject.transform.SetParent(PlayerAvatarInstance.transform, false);
-        valuableObject.gameObject.AddComponent<Rigidbody>();
+        ValuableComponent.durabilityPreset = ScriptableObject.CreateInstance<Durability>();
+        ValuableComponent.durabilityPreset.durability = 999;
+        ValuableComponent.durabilityPreset.fragility = 0;
+        ValuableComponent.physAttributePreset = ScriptableObject.CreateInstance<PhysAttribute>();
+        ValuableComponent.transform.localScale = Vector3.zero;
+        ValuableComponent.transform.SetParent(player.transform, false);
+        ValuableComponent.gameObject.AddComponent<Rigidbody>();
 
-        var ptv = valuableObject.gameObject.AddComponent<PhotonTransformView>();
-        var pgoid = valuableObject.gameObject.AddComponent<PhysGrabObjectImpactDetector>();
-        var pgo = valuableObject.gameObject.AddComponent<PhysGrabObject>();
-        var pv = valuableObject.gameObject.AddComponent<PhotonView>();
+        var ptv = ValuableComponent.gameObject.AddComponent<PhotonTransformView>();
+        var pgoid = ValuableComponent.gameObject.AddComponent<PhysGrabObjectImpactDetector>();
+        var pgo = ValuableComponent.gameObject.AddComponent<PhysGrabObject>();
+        var pv = ValuableComponent.gameObject.AddComponent<PhotonView>();
 
         pv.ObservedComponents = [pgo, pgoid, ptv];
         pgoid.enabled = false;
         pgo.enabled = false;
         pgo.transform.localScale = Vector3.zero;
-        pgo.transform.SetParent(valuableObject.transform, false);
+        pgo.transform.SetParent(ValuableComponent.transform, false);
         pgoid.transform.localScale = Vector3.zero;
-        pgoid.transform.SetParent(valuableObject.transform, false);
+        pgoid.transform.SetParent(ValuableComponent.transform, false);
 
-        var rvc = valuableObject.gameObject.AddComponent<RoomVolumeCheck>();
+        var rvc = ValuableComponent.gameObject.AddComponent<RoomVolumeCheck>();
         rvc.CurrentRooms = [];
 
-        valuableObject.particleColors = new Gradient {
-                                                         alphaKeys = [new GradientAlphaKey(1, 0)],
-                                                         colorKeys = [new GradientColorKey(Color.yellow, 0)]
-                                                     };
+        ValuableComponent.particleColors = new Gradient {
+                                                            alphaKeys = [new GradientAlphaKey(1, 0)],
+                                                            colorKeys = [new GradientColorKey(Color.yellow, 0)]
+                                                        };
 
-        var physAttributePresetRef = FieldRefAccess<ValuableObject, PhysAttribute>("physAttributePreset");
-        physAttributePresetRef.Invoke(valuableObject) = ScriptableObject.CreateInstance<PhysAttribute>();
+        _physAttributePresetRef.Invoke(ValuableComponent) = ScriptableObject.CreateInstance<PhysAttribute>();
 
         lastHealth = lastLevel = -1;
     }
@@ -118,19 +127,18 @@ public class GoldenHeart : MonoBehaviour {
         if (!SemiFunc.IsMasterClientOrSingleplayer()) return;
         if (!Traverse.Create(RoundDirector.instance).Field("extractionPointsFetched").GetValue<bool>()) return;
 
-        var healthRef = FieldRefAccess<PlayerHealth, int>("health");
         var heartOfGoldUpgrade = SLRUpgradePack.HeartOfGoldUpgradeInstance;
 
         if (!heartOfGoldUpgrade.UpgradeEnabled.Value) return;
         if (Pause) return;
-        if (heartOfGoldUpgrade.UpgradeRegister.GetLevel(PlayerAvatarInstance) == 0 || healthRef.Invoke(PlayerAvatarInstance.playerHealth) == 0) return;
-        if (lastLevel == heartOfGoldUpgrade.UpgradeRegister.GetLevel(PlayerAvatarInstance) &&
-            lastHealth == healthRef.Invoke(PlayerAvatarInstance.playerHealth)) return;
+        if (heartOfGoldUpgrade.UpgradeRegister.GetLevel(player) == 0 || _healthRef.Invoke(player.playerHealth) == 0) return;
+        if (lastLevel == heartOfGoldUpgrade.UpgradeRegister.GetLevel(player) &&
+            lastHealth == _healthRef.Invoke(player.playerHealth)) return;
 
-        lastLevel = heartOfGoldUpgrade.UpgradeRegister.GetLevel(PlayerAvatarInstance);
-        lastHealth = healthRef.Invoke(PlayerAvatarInstance.playerHealth);
+        lastLevel = heartOfGoldUpgrade.UpgradeRegister.GetLevel(player);
+        lastHealth = _healthRef.Invoke(player.playerHealth);
 
-        if (GetComponent<ValuableObject>() == null || !GetComponent<ValuableObject>()) {
+        if (ValuableComponent == null || !ValuableComponent) {
             CreateOnlyMe();
         }
 
@@ -140,19 +148,16 @@ public class GoldenHeart : MonoBehaviour {
 
     [PunRPC]
     private void UpdateOnlyMeRPC() {
-        var healthRef = FieldRefAccess<PlayerHealth, int>("health");
         var heartOfGoldUpgrade = SLRUpgradePack.HeartOfGoldUpgradeInstance;
 
-        if (!FieldRefAccess<ValuableObject, bool>("discovered").Invoke(GetComponent<ValuableObject>()))
-            GetComponent<ValuableObject>().Discover(ValuableDiscoverGraphic.State.Discover);
+        if (!_discoveredRef.Invoke(ValuableComponent))
+            ValuableComponent.Discover(ValuableDiscoverGraphic.State.Discover);
 
-        var dollarValueCurrentRef = FieldRefAccess<ValuableObject, float>("dollarValueCurrent");
+        _dollarValueCurrentRef.Invoke(ValuableComponent) =
+            heartOfGoldUpgrade.Calculate(_healthRef.Invoke(player.playerHealth) * heartOfGoldUpgrade.BaseHeartValue.Value,
+                                         player, heartOfGoldUpgrade.UpgradeRegister.GetLevel(player));
 
-        dollarValueCurrentRef.Invoke(GetComponent<ValuableObject>()) =
-            heartOfGoldUpgrade.Calculate(healthRef.Invoke(PlayerAvatarInstance.playerHealth) * heartOfGoldUpgrade.BaseHeartValue.Value,
-                                         PlayerAvatarInstance, heartOfGoldUpgrade.UpgradeRegister.GetLevel(PlayerAvatarInstance));
-
-        ValuableObjectValuePatch.Action(GetComponent<ValuableObject>());
+        ValuableObjectValuePatch.Action(ValuableComponent);
     }
 
     public void PauseLogic(bool pause) {
@@ -168,6 +173,8 @@ public class GoldenHeart : MonoBehaviour {
 
 [HarmonyPatch(typeof(ExtractionPoint))]
 public class ExrtractionPointdestoryPatch {
+    private static FieldRef<RoundDirector, int>? _totalHaulRef = FieldRefAccess<RoundDirector, int>("totalHaul");
+
     [HarmonyPatch("DestroyAllPhysObjectsInHaulList")]
     [HarmonyPrefix]
     private static bool DestroyAllPrefix() {
@@ -175,18 +182,19 @@ public class ExrtractionPointdestoryPatch {
         if (!SemiFunc.IsMasterClientOrSingleplayer() || !heartOfGoldUpgrade.UpgradeEnabled.Value)
             return true;
 
-        var totalHaulRef = FieldRefAccess<RoundDirector, int>("totalHaul");
-
         foreach (GameObject dollarHaul in RoundDirector.instance.dollarHaulList) {
             if (dollarHaul && dollarHaul.GetComponent<PhysGrabObject>()) {
-                totalHaulRef.Invoke(RoundDirector.instance) += (int)Traverse
-                                                                   .Create(dollarHaul.GetComponent<ValuableObject>())
-                                                                   .Field("dollarValueCurrent").GetValue<float>();
+                _totalHaulRef.Invoke(RoundDirector.instance) += (int)Traverse
+                                                                    .Create(dollarHaul.GetComponent<ValuableObject>())
+                                                                    .Field("dollarValueCurrent").GetValue<float>();
 
-                if (dollarHaul.TryGetComponent<GoldenHeart>(out var heart)) {
-                    SLRUpgradePack.Logger.LogInfo("Player in extraction zone counts as valuable");
-                    heart.PauseLogic(true);
-                    heart.DestroyOnlyMe();
+                if (dollarHaul.TryGetComponent<ValuableObject>(out var valuableObject) && valuableObject.name.Equals("Health Grab")) {
+                    foreach (var idHeart in heartOfGoldUpgrade.GoldenHearts)
+                        if (idHeart.Value.ValuableComponent == valuableObject) {
+                            SLRUpgradePack.Logger.LogInfo($"Player {SemiFunc.PlayerGetName(idHeart.Value.player)} in extraction zone counts as valuable");
+                            idHeart.Value.PauseLogic(true);
+                            idHeart.Value.DestroyOnlyMe();
+                        }
                 } else
                     dollarHaul.GetComponent<PhysGrabObject>().DestroyPhysGrabObject();
             }
@@ -194,11 +202,10 @@ public class ExrtractionPointdestoryPatch {
 
         foreach (PlayerAvatar player in GameDirector.instance.PlayerList) {
             player.playerDeathHead.Revive();
+        }
 
-            if (player.TryGetComponent<GoldenHeart>(out var heart)) {
-                heart.PauseLogic(false);
-                heart.lastHealth = heart.lastLevel = -1;
-            }
+        foreach (var idHeart in heartOfGoldUpgrade.GoldenHearts) {
+            idHeart.Value.PauseLogic(false);
         }
 
         return false;
@@ -215,22 +222,18 @@ public class ExrtractionPointdestoryPatch {
             !heartOfGoldUpgrade.UpgradeEnabled.Value)
             return true;
 
-        var totalHaulRef = FieldRefAccess<RoundDirector, int>("totalHaul");
+        _totalHaulRef.Invoke(RoundDirector.instance) += (int)Traverse
+                                                            .Create(RoundDirector.instance.dollarHaulList[0]
+                                                                                 .GetComponent<ValuableObject>())
+                                                            .Field("dollarValueCurrent").GetValue<float>();
 
-        totalHaulRef.Invoke(RoundDirector.instance) += (int)Traverse
-                                                           .Create(RoundDirector.instance.dollarHaulList[0]
-                                                                                .GetComponent<ValuableObject>())
-                                                           .Field("dollarValueCurrent").GetValue<float>();
-
-        if (RoundDirector.instance.dollarHaulList[0].TryGetComponent<GoldenHeart>(out var heart)) {
-            SLRUpgradePack.Logger.LogInfo("Player in extraction zone counts as valuable");
-            heart.PauseLogic(true);
-            heart.DestroyOnlyMe();
-
-            if (RoundDirector.instance.dollarHaulList[0].TryGetComponent<ExtraLife>(out var extraLife)) {
-                SLRUpgradePack.Logger.LogInfo("Pausing extra life logic for player in extraction zone");
-                extraLife.PauseLogic(true);
-            }
+        if (RoundDirector.instance.dollarHaulList[0].TryGetComponent<ValuableObject>(out var valuableObject) && valuableObject.name.Equals("Health Grab")) {
+            foreach (var idHeart in heartOfGoldUpgrade.GoldenHearts)
+                if (idHeart.Value.ValuableComponent == valuableObject) {
+                    SLRUpgradePack.Logger.LogInfo($"Player {SemiFunc.PlayerGetName(idHeart.Value.player)} in extraction zone counts as valuable");
+                    idHeart.Value.PauseLogic(true);
+                    idHeart.Value.DestroyOnlyMe();
+                }
         } else
             RoundDirector.instance.dollarHaulList[0].GetComponent<PhysGrabObject>().DestroyPhysGrabObject();
 
@@ -243,6 +246,6 @@ public class ExrtractionPointdestoryPatch {
 [HarmonyPatch(typeof(RoundDirector), "Update")]
 public class RoundDirectorUpdatePatch {
     private static void Prefix(RoundDirector __instance) {
-        __instance.dollarHaulList.RemoveAll(go => go == null || !go);
+        __instance.dollarHaulList.RemoveAll(go => go == null || !go.TryGetComponent<ValuableObject>(out var valuableObject) || !valuableObject);
     }
 }